🎨 fix: move out gql files
