+++ src/api/notice/notice-gql.ts
@@ -1,0 +1,31 @@
+import { gql } from '@/generated';
+
+export const GET_NOTICES = gql(`
+  query GetNotices($offset: Int, $limit: Int) {
+    notices(offset: $offset, limit: $limit) {
+      list {
+        id
+        title
+        views
+        body
+        deadline
+        createdAt
+        author
+        imageUrl
+        tags {
+          id
+          name
+        }
+      }
+      total
+    }
+  }
+`);
+
+export const CREATE_NOTICE = gql(`
+  mutation CreateNotice($title: String!, $body: String!, $deadline: Date, $tags: [Int!], $images: [String!]) {
+    createNotice(title: $title, body: $body, deadline: $deadline, tags: $tags, images: $images) {
+      id
+    }
+  }
+`);
