+++ src/api/notice/notice.ts
@@ -73,33 +73,3 @@ export const getNotice = async (id: number) =>
     ...data,
     deadline: data.deadline ? data.deadline : undefined,
   }));
-
-export const GET_NOTICES = gql(`
-  query GetNotices($offset: Int, $limit: Int) {
-    notices(offset: $offset, limit: $limit) {
-      list {
-        id
-        title
-        views
-        body
-        deadline
-        createdAt
-        author
-        imageUrl
-        tags {
-          id
-          name
-        }
-      }
-      total
-    }
-  }
-`);
-
-export const CREATE_NOTICE = gql(`
-  mutation CreateNotice($title: String!, $body: String!, $deadline: Date, $tags: [Int!], $images: [String!]) {
-    createNotice(title: $title, body: $body, deadline: $deadline, tags: $tags, images: $images) {
-      id
-    }
-  }
-`);
